lab-04

+++ Lab-01/Controllers/ProductController.cs
@@ -4,6 +4,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Lab_01.Models;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace Lab_01.Controllers
 {
@@ -17,7 +19,7 @@ namespace Lab_01.Controllers
         }
         public IActionResult Index()
         {
-            var product = _context.Products.ToList();
+            var product = _context.Products.Include(p => p.Category).ToList();
             return View(product);
         }
 
@@ -25,6 +27,7 @@ namespace Lab_01.Controllers
         //trả về danh sách sản phẩm
         public IActionResult Create()
         {
+            ViewBag.Categories = new SelectList(_context.Categories, "Id", "Name");
             return View();
         }
 
@@ -35,21 +38,49 @@ namespace Lab_01.Controllers
         {
             if (ModelState.IsValid)
             {
-
                 _context.Products.Add(product);
                 _context.SaveChanges();
-                
+                return RedirectToAction("Index");
             }
-            return RedirectToAction("Index");
+            ViewBag.Categories = new SelectList(_context.Categories, "Id", "Name", product.CategoryId);
+            return View(product);
         }
 
-
-        public IActionResult Delete(int id)
+        public IActionResult Edit(int id)
         {
-            var product = _context.Products.Find(id); 
+            var product = _context.Products.Find(id);
+            if (product == null) return NotFound();
+
+            ViewBag.Categories = new SelectList(_context.Categories, "Id", "Name", product.CategoryId);
+            return View(product);
+        }
+
+        // POST: Sửa sản phẩm
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public IActionResult Edit(ProductModel product)
+        {
+            if (ModelState.IsValid)
+            {
+                _context.Products.Update(product);
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            ViewBag.Categories = new SelectList(_context.Categories, "Id", "Name", product.CategoryId);
+            return View(product);
+        }
+
+        // POST: Xóa sản phẩm
+        [HttpPost, ActionName("Delete")]
+        [ValidateAntiForgeryToken]
+        public IActionResult DeleteConfirmed(int id)
+        {
+            var product = _context.Products.Find(id);
+            if (product != null)
+            {
                 _context.Products.Remove(product);
                 _context.SaveChanges();
-
+            }
             return RedirectToAction("Index");
         }
     }
